OPIK-744 post rebase fixes
OpenaiModelName class was moved - switch to use the new class
